Eliminate compiler warnings

+++ MagickCore/magick-type.h
@@ -181,7 +181,7 @@ typedef enum
 #  define IsNaN(a) ((a) != (a))
 #endif
 #if !defined(INFINITY)
-#  define INFINITY (-logf(0f))
+#  define INFINITY ((double) -logf(0f))
 #endif
 
 typedef struct _BlobInfo BlobInfo;
